Refactor logging utils
