+++ src/prime_rl/eval/eval.py
@@ -9,7 +9,7 @@ from prime_rl.orchestrator.client import (
     setup_client,
     update_weights,
 )
-from prime_rl.orchestrator.logger import setup_logger
+from prime_rl.utils.logger import setup_logger
 from prime_rl.utils.monitor import setup_monitor
 from prime_rl.utils.pydantic_config import parse_argv
 from prime_rl.utils.utils import clean_exit
@@ -18,7 +18,9 @@ from prime_rl.utils.utils import clean_exit
 @clean_exit
 async def eval(config: OfflineEvalConfig):
     # Initialize the logger
-    logger = setup_logger(config.log, config.output_dir)
+    logger = setup_logger(
+        config.log.level, log_file=config.output_dir / "logs" / "eval.log" if config.log.file else None
+    )
     logger.info("Starting evaluation")
     logger.info(f"Model: {config.model}")
     logger.info(f"Sampling: {config.sampling}")
