+++ src/prime_rl/orchestrator/orchestrator.py
@@ -24,7 +24,7 @@ from prime_rl.orchestrator.client import (
 from prime_rl.orchestrator.config import OrchestratorConfig
 from prime_rl.orchestrator.buffer import setup_buffer, make_rollouts, Rollout
 from prime_rl.orchestrator.batch import prepare_batch
-from prime_rl.orchestrator.logger import setup_logger
+from prime_rl.utils.logger import setup_logger
 from prime_rl.orchestrator.advantage import compute_advantages
 from prime_rl.orchestrator.utils import (
     wait_for_weight_checkpoint,
@@ -48,7 +48,9 @@ import numpy as np
 @logger.catch(reraise=True)
 async def orchestrate(config: OrchestratorConfig):
     # Initialize the logger
-    logger = setup_logger(config.log, config.output_dir)
+    logger = setup_logger(
+        config.log.level, log_file=config.output_dir / "logs" / "orchestrator.log" if config.log.file else None
+    )
     logger.info("Starting orchestrator")
 
     # Print warning if running in benchmark mode
