+++ src/prime_rl/utils/logger.py
@@ -1,25 +1,17 @@
 import sys
 from pathlib import Path
 
-from loguru._logger import Logger
+from loguru import logger
 
-from prime_rl.utils.config import LogConfig
-
-# Global loguru logger instance
-_LOGGER: Logger | None = None
+# Global logger instance
+_LOGGER = None
 
 NO_BOLD = "\033[22m"
 RESET = "\033[0m"
 
 
-def format_time(config: LogConfig) -> str:
-    time = "<dim>{time:HH:mm:ss}</dim>"
-    if config.utc:
-        time = "<dim>{time:zz HH:mm:ss!UTC}</dim>"
-    return time
-
-
-def format_message() -> str:
+def setup_logger(log_level: str, log_console: bool = True, log_file: Path | None = None):
+    # Format message
     message = "".join(
         [
             " <level>{level: >7}</level>",
@@ -28,27 +20,24 @@ def format_message() -> str:
             f"{RESET}</level>",
         ]
     )
-    return message
+    time = "<dim>{time:HH:mm:ss}</dim>"
+    if log_level.upper() != "DEBUG":
+        debug = ""
+    else:
+        debug = "".join([f"<level>{NO_BOLD}", " [{file}::{line}]", f"{RESET}</level>"])
+    format = time + message + debug
 
-
-def format_debug(config: LogConfig) -> str:
-    if config.level.upper() != "DEBUG":
-        return ""
-    return "".join([f"<level>{NO_BOLD}", " [{file}::{line}]", f"{RESET}</level>"])
-
-
-def setup_handlers(logger: Logger, format: str, config: LogConfig, rank: int, output_dir: Path) -> Logger:
     # Remove all default handlers
     logger.remove()
 
     # Install console on the master rank
-    if rank == 0:
-        logger.add(sys.stdout, format=format, level=config.level.upper(), colorize=True)
+    if log_console:
+        logger.add(sys.stdout, format=format, level=log_level.upper(), colorize=True)
 
     # If specified, install file handlers on all ranks
-    if config.file:
-        log_file = output_dir / "logs" / f"rank_{rank}.log"
-        logger.add(log_file, format=format, level=config.level.upper())
+    if log_file is not None:
+        # We do not colorize so that the file log is easier to read
+        logger.add(log_file, format=format, level=log_level.upper())
 
     # Disable critical logging
     logger.critical = lambda _: None
@@ -56,17 +45,7 @@ def setup_handlers(logger: Logger, format: str, config: LogConfig, rank: int, ou
     return logger
 
 
-def set_logger(logger: Logger) -> None:
-    """
-    Set the global logger. This function is shared across submodules such as
-    training and inference, and should be called *exactly once* from a
-    module-specific `setup_logger` function with the logger instance.
-    """
-    global _LOGGER
-    _LOGGER = logger
-
-
-def get_logger() -> Logger:
+def get_logger():
     """
     Get the global logger. This function is shared across submodules such as
     training and inference to accesst the global logger instance. Raises if the
@@ -81,7 +60,7 @@ def get_logger() -> Logger:
     return _LOGGER
 
 
-def reset_logger() -> None:
+def reset_logger():
     """Reset the global logger. Useful mainly in test to clear loggers between tests."""
     global _LOGGER
     _LOGGER = None
